refactor: split into different files
Fixes #103
