+++ .gitignore
@@ -14,6 +14,9 @@
 *.d.ts.map
 *.d.mts
 *.d.mts.map
+!/lib/**/*-types.d.ts
+!/index.d.ts
+!/index.d.mts
 
 # Library specific ones
-index.mjs
+/lib/render.mjs
