+++ index.d.ts
@@ -1,0 +1,5 @@
+export type * from './lib/element-types.d.ts';
+
+export { h, html, rawHtml } from './lib/htm.js';
+export { render, renderToString } from './lib/render.js';
+export { generatorToString } from './lib/utils.js';
