+++ lib/util-types.d.ts
@@ -1,0 +1,2 @@
+export type IterableIteratorMaybeAsync<T> = T[] | IterableIterator<T> | AsyncIterableIterator<T>;
+export type MaybeArray<T> = T | T[];
