+++ package.json
@@ -45,14 +45,20 @@
     "index.d.ts.map",
     "index.d.ts",
     "index.js",
-    "index.mjs"
+    "index.mjs",
+    "lib/**/*.d.mts.map",
+    "lib/**/*.d.mts",
+    "lib/**/*.d.ts.map",
+    "lib/**/*.d.ts",
+    "lib/**/*.js",
+    "lib/**/*.mjs"
   ],
   "scripts": {
     "build:0-clean": "run-s clean",
-    "build:1-esm": "linemod -e mjs index.js",
+    "build:1-esm": "linemod -e mjs lib/render.js",
     "build:2-declaration": "tsc -p declaration.tsconfig.json",
     "build": "run-s build:*",
-    "check:0": "run-s clean",
+    "check:0": "run-s clean build:1-esm",
     "check:1:installed-check": "installed-check",
     "check:1:knip": "knip",
     "check:1:lint": "eslint --report-unused-disable-directives .",
@@ -63,8 +69,10 @@
     "check:2": "run-s build",
     "check:3": "tsd",
     "check": "run-s check:*",
-    "clean:declarations": "rm -rf $(find . -maxdepth 2 -type f -name '*.d.ts*')",
-    "clean:esm-declarations": "rm -rf $(find . -maxdepth 2 -type f -name '*.d.mts*')",
+    "clean:declarations-top": "rm -rf $(find . -maxdepth 1 -type f -name '*.d.ts*' ! -name 'index.d.ts')",
+    "clean:declarations-lib": "rm -rf $(find lib -type f -name '*.d.ts*' ! -name '*-types.d.ts')",
+    "clean:esm-declarations-top": "rm -rf $(find . -maxdepth 1 -type f -name '*.d.mts*' ! -name 'index.d.mts')",
+    "clean:esm-declarations-lib": "rm -rf $(find lib -type f -name '*.d.mts*')",
     "clean": "run-p clean:*",
     "prepare": "husky",
     "prepublishOnly": "run-s build",
@@ -97,7 +105,7 @@
     "husky": "^9.1.1",
     "installed-check": "^9.3.0",
     "knip": "^5.26.0",
-    "linemod": "^2.0.0",
+    "linemod": "^2.0.1",
     "mocha": "^10.6.0",
     "nodemon": "^3.1.4",
     "npm-run-all2": "^6.2.2",
